src/requirements.txt
aiodns==2.0.0; python_version >= "3.6"
aiohttp-retry==2.3.5
aiohttp==3.7.4.post0; python_version >= "3.6"
aiokafka==0.7.0; python_version >= "3.7"
asgiref==3.3.4; python_version >= "3.6"
async-timeout==3.0.1; python_full_version >= "3.5.3" and python_version >= "3.6"
attrs==20.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
beautifulsoup4==4.9.3
broadcaster-noteable==0.2.1; python_version >= "3.7"
brotlipy==0.7.0; python_version >= "3.6"
bs4==0.0.1
cached-property==1.5.2; python_version >= "3.7" and python_version < "4.0"
cchardet==2.1.7; python_version >= "3.6"
certifi==2020.12.5; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
cffi==1.14.5; python_version >= "3.6"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
click==7.1.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
colorama==0.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
crc32c==2.2
dawn==0.11.0
django-anymail==8.2; python_version >= "3.5"
django-cors-headers==3.7.0; python_version >= "3.6"
django-extensions==3.1.2; python_version >= "3.6"
django-filter==2.4.0; python_version >= "3.5"
django-k8s==0.2.9
django-prometheus==2.1.0
django-registration-redux==2.9
django-request-logging==0.7.2
django-upload-validator==1.1.5
django-widget-tweaks==1.4.8
django==3.2; python_version >= "3.6"
djangorestframework-simplejwt==4.6.0; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
docopt==0.6.2; python_version >= "3.4"
ebooklib==0.17.1
graphql-core==3.1.4; python_version >= "3.7" and python_version < "4"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.12.0; python_version >= "3.6"
html5lib==1.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
httptools==0.1.1; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
hupper==1.10.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.4.0"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
kafka-python==2.0.2
libgravatar==0.2.5
lxml==4.6.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
multidict==5.1.0; python_version >= "3.6"
opencc==1.1.1.post1
prometheus-client==0.10.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pycares==3.1.1; python_version >= "3.6"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pygments==2.8.1; python_version >= "3.7" and python_version < "4.0"
pyjwt==2.0.1; python_version >= "3.7"
python-dateutil==2.8.1; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
python-dotenv==0.17.0
python-magic==0.4.22; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
python-multipart==0.0.5; python_version >= "3.7" and python_version < "4.0"
pytz==2021.1; python_version >= "3.6"
pyyaml==5.4.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
requests==2.25.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
six==1.15.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
soupsieve==2.2.1; python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
starlette==0.14.2; python_version >= "3.7" and python_version < "4.0"
strawberry-graphql-django==0.0.6; python_version >= "3.7" and python_version < "4.0"
strawberry-graphql==0.47.1; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.7.4.3; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.4; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.5"
uvicorn==0.13.4
uvloop==0.15.2; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.7; python_version >= "3.5"
webencodings==0.5.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
websockets==8.1; python_full_version >= "3.6.1"
webvtt-py==0.4.6; python_version >= "3.4"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
xmltodict==0.12.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
yarl==1.6.3; python_version >= "3.6"
